refactor(app): move transfer-details
