+++ app/src/lib/components/transfer-details.svelte
