+++ app/src/routes/explorer/transfers/[source]/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 import ChainsGate from "$lib/components/chains-gate.svelte"
-import TransferDetails from "./(components)/transfer-details.svelte"
+import TransferDetails from "$lib/components/transfer-details.svelte"
 </script>
 
 <ChainsGate let:chains>
